support ring attention for sft&dpo (varlen only, llm only)
