+++ swift/trainers/sequence_parallel/ulysses.py
@@ -18,87 +18,15 @@ from trl.extras.profiling import profiling_decorator
 
 from swift.llm import DataLoaderDispatcher, DataLoaderShard, get_llm_model, to_device
 from swift.utils import get_current_device, get_device, get_dist_setting, seed_worker
-from .base import SequenceParallel
+from .base import CommonSequenceParallel
+from .utils import (
+    GatherLoss, ChunkedCrossEntropyLoss, loss_scale_sp_func, _prepare_inputs,
+    SequenceParallelSampler, SequenceParallelDispatcher, setup_compute_acc, get_common_dataloader,
+    get_per_token_logps
+)
 
 assert version.parse(torch.__version__) >= version.parse('2.0.0')
 torch._dynamo.config.capture_dynamic_output_shape_ops = True
-
-
-class GatherLoss(torch.autograd.Function):
-    """Gather loss from sequence group"""
-
-    @staticmethod
-    def forward(ctx, loss, labels, process_group, gather_idx=None):
-        """
-        Args:
-            loss: loss tensor after splitting
-            labels: labels tensor after splitting
-            process_group: the sequence parallel group
-            gather_idx: gather the tensors on this dim
-        """
-        ctx.process_group = process_group
-        shape0 = labels.shape[0]
-        ctx.scatter_shape = labels.shape[gather_idx or 0]
-        ctx.gather_idx = gather_idx or 0
-        world_size = dist.get_world_size(group=process_group)  # the sp world size
-        output = torch.empty((shape0 * world_size, *loss.shape[1:]), dtype=loss.dtype, device=loss.device)
-        # gather all from sp group
-        dist.all_gather_into_tensor(output, loss, group=process_group)
-        if gather_idx is not None:
-            output = torch.cat(output.split(shape0, dim=0), dim=gather_idx)
-        labels_output = torch.empty((shape0 * world_size, *labels.shape[1:]), dtype=labels.dtype, device=labels.device)
-        dist.all_gather_into_tensor(labels_output, labels, group=process_group)
-        if gather_idx is not None:
-            labels_output = torch.cat(labels_output.split(shape0, dim=0), dim=gather_idx)
-        return output, labels_output
-
-    @staticmethod
-    def backward(ctx, *grad_output):
-        _grad = grad_output[0] * dist.get_world_size(group=ctx.process_group)
-        return _grad.split(
-            ctx.scatter_shape, dim=ctx.gather_idx)[dist.get_rank(ctx.process_group)].contiguous(), None, None, None
-
-
-class ChunkedCrossEntropyLoss(torch.autograd.Function):
-
-    @staticmethod
-    def forward(ctx, logits, labels, chunk_size):
-        ctx.save_for_backward(logits, labels)
-        ctx.chunk_size = chunk_size
-
-        losses = []
-        for i in range(math.ceil(logits.shape[0] / chunk_size)):
-            l_start = i * chunk_size
-            l_end = min((i + 1) * chunk_size, logits.shape[0])
-            logits_chunk = logits[l_start:l_end]
-            labels_chunk = labels[l_start:l_end]
-            loss_fct = CrossEntropyLoss(reduction='none')
-            loss_chunk = loss_fct(logits_chunk, labels_chunk)
-            losses.append(loss_chunk)
-            del logits_chunk
-            del labels_chunk
-        all_losses = torch.cat(losses)
-        return all_losses
-
-    @staticmethod
-    def backward(ctx: Any, *grad_outputs: Any):
-        logits, labels = ctx.saved_tensors
-        chunk_size = ctx.chunk_size
-
-        for i in range(math.ceil(logits.shape[0] / chunk_size)):
-            l_start = i * chunk_size
-            l_end = min((i + 1) * chunk_size, logits.shape[0])
-            logits_chunk = logits[l_start:l_end].detach().requires_grad_(True)
-            labels_chunk = labels[l_start:l_end]
-            loss_fct = CrossEntropyLoss(reduction='none')
-            with torch.enable_grad():
-                loss_chunk = loss_fct(logits_chunk, labels_chunk)
-                grad_output_chunk = grad_outputs[0][l_start:l_end]
-                _loss_chunk = (loss_chunk * grad_output_chunk).sum()
-                grad_chunk = torch.autograd.grad(_loss_chunk, logits_chunk, retain_graph=False)[0]
-                logits[l_start:l_end] = grad_chunk
-
-        return logits, None, None
 
 
 torch_compile_options = {
@@ -108,38 +36,6 @@ torch_compile_options = {
     'trace.enabled': False,
     'triton.cudagraphs': False,
 }
-
-
-# TODO not work with `ChunkedCrossEntropyLoss.apply`
-# @torch.compile(dynamic=True, fullgraph=True, options=torch_compile_options)
-def loss_scale_sp_func(outputs, labels, loss_scale=None, num_items_in_batch=None, ulysses=None) -> torch.Tensor:
-    if hasattr(outputs, 'logits'):
-        logits = outputs.logits
-    else:
-        logits = outputs
-    device = logits.device
-    if labels.shape[1] > logits.shape[1]:
-        _, _, labels, _, _, loss_scale = ulysses.pad_and_split_inputs(None, None, labels, None, None, loss_scale)
-    logits = logits.view(-1, logits.shape[-1])
-
-    labels = labels.flatten().to(device)
-    sploss_parallel_size = int(os.environ.get('CELOSS_PARALLEL_SIZE', '0'))
-    if sploss_parallel_size > 0:
-        loss = ChunkedCrossEntropyLoss.apply(logits, labels, sploss_parallel_size)
-    else:
-        loss_fct = CrossEntropyLoss(reduction='none')
-        loss = loss_fct(logits, labels)
-    if loss_scale is not None:
-        loss_scale = loss_scale.flatten().to(device)
-        loss = (loss_scale * loss)
-    loss, labels = GatherLoss.apply(loss, labels, ulysses.sp_group)
-    mask = (labels != -100)
-    total_loss = loss[mask].sum()
-    if num_items_in_batch is None:
-        total_loss = total_loss / mask.sum()
-    else:
-        total_loss = total_loss / num_items_in_batch
-    return total_loss
 
 
 @profiling_decorator
@@ -160,42 +56,12 @@ def _prepare_inputs_grpo(self, generation_batch):
     return inputs
 
 
-def _prepare_inputs(self, inputs, ulysses):
-    if 'labels' in inputs:
-        labels = inputs['labels']
-        _, _, labels, _, _, _ = ulysses.pad_and_split_inputs(None, None, labels, None, None, None)
-        inputs['labels'] = labels
-    return self._origin_prepare_inputs(inputs)
-
-
 def old_policy(self):
     ulysses = self.ulysses
     # changes: `* ulysses.sp_world_size`
     return (self.num_iterations > 1
             or self.args.steps_per_generation * ulysses.sp_world_size > self.args.gradient_accumulation_steps)
 
-
-# For DPO
-def get_per_token_logps(logits: torch.FloatTensor,
-                        labels: torch.LongTensor,
-                        label_pad_token_id=-100,
-                        ulysses=None) -> Tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    if labels.shape[1] > logits.shape[1]:
-        _, _, labels, _, _, _ = ulysses.pad_and_split_inputs(None, None, labels, None, None, None)
-    loss_mask = labels != label_pad_token_id
-    labels = labels.clone()  # No need to shift, pad and split has shifted the inputs.
-    labels[~loss_mask] = 0
-    labels = labels.to(logits.device)
-    loss_mask = loss_mask.to(logits.device)
-    mean_logits = logits.mean(-1)
-    per_token_logps = torch.gather(logits.log_softmax(-1), dim=2, index=labels.unsqueeze(2)).squeeze(2)
-    total_per_token_logps, total_loss_mask = GatherLoss.apply(per_token_logps, loss_mask, ulysses.sp_group, 1)
-
-    world_size = dist.get_world_size(group=ulysses.sp_group)
-    total_mean_logits = mean_logits.new_empty((mean_logits.shape[0], mean_logits.shape[1] * world_size))
-    dist.all_gather_into_tensor(total_mean_logits, mean_logits, group=ulysses.sp_group)
-    total_per_token_logps[~total_loss_mask] = 0
-    return total_per_token_logps, total_mean_logits, total_loss_mask
 
 
 @contextmanager
@@ -376,84 +242,6 @@ def split_by_mini_batches(self, inputs, advantages):
     return spg_chunks, advantage_chunks
 
 
-class UlyssesSampler(Sampler):
-
-    # Code borrowed from mmengine
-    def __init__(self, ulysses, dataset, shuffle: bool = True, seed=None, round_up: bool = True) -> None:
-        self.ulysses = ulysses
-        rank = dist.get_rank(ulysses.device_mesh['data'].get_group())
-        world_size = ulysses.device_mesh['data'].size()
-        self.rank = rank
-        self.world_size = world_size
-
-        self.dataset = dataset
-        self.shuffle = shuffle
-        assert seed is not None
-        self.seed = seed
-        self.epoch = 0
-        self.round_up = round_up
-
-        if self.round_up:
-            self.num_samples = math.ceil(len(self.dataset) / world_size)
-            self.total_size = self.num_samples * self.world_size
-        else:
-            self.num_samples = math.ceil((len(self.dataset) - rank) / world_size)
-            self.total_size = len(self.dataset)
-
-    def __iter__(self) -> Iterator[int]:
-        if self.shuffle:
-            g = torch.Generator()
-            g.manual_seed(self.seed + self.epoch)
-            indices = torch.randperm(len(self.dataset), generator=g).tolist()
-        else:
-            indices = torch.arange(len(self.dataset)).tolist()
-
-        if self.round_up:
-            indices = (indices * int(self.total_size / len(indices) + 1))[:self.total_size]
-
-        indices = indices[self.rank:self.total_size:self.world_size]
-        return iter(indices)
-
-    def __len__(self) -> int:
-        return self.num_samples
-
-    def set_epoch(self, epoch: int) -> None:
-        self.epoch = epoch
-
-
-class UlyssesDispatcher(DataLoaderDispatcher):
-
-    def __init__(self, base_dataloader, ulysses, device=None):
-        super().__init__(base_dataloader)
-        self.ulysses = ulysses
-        self.device = device
-
-    def _scatter_object_list(self, inputs):
-        if not dist.is_initialized():
-            return inputs[0]
-        outputs = [None]
-        global_src_rank = dist.get_global_rank(self.ulysses.dp_group, 0)
-        dist.scatter_object_list(outputs, inputs, global_src_rank, group=self.ulysses.dp_group)
-        return outputs[0]
-
-    def __iter__(self):
-        base_iter = iter(self.base_dataloader)
-        while True:
-            if self.ulysses.dp_rank == 0:
-                try:
-                    data = [next(base_iter) for _ in range(self.ulysses.dp_world_size)]
-                except StopIteration:
-                    data = [None] * self.ulysses.dp_world_size
-                data = self._scatter_object_list(data)
-            else:
-                data = self._scatter_object_list(None)
-            if data is None:
-                break
-            if self.device:
-                data = to_device(data, self.device)
-            yield data
-
-
 # Code borrowed from deepspeed, here is why:
 # 1. Reduce the dependency
 # 2. The original code is complex
@@ -576,26 +364,19 @@ class DistributedAttention(torch.nn.Module):
         return output
 
 
-class Ulysses(SequenceParallel):
+class Ulysses(CommonSequenceParallel):
 
     def __init__(self):
+        super().__init__()
         self.split_in_forward = None
-        self.dp_world_size = None
-        self.sp_world_size = None
-        self.model_dtype = None
         self.causal_mask_func = None
-        self.device_mesh = None
-        self._inited = False
 
     def init_sequence_parallel(self, size):
         if self._inited:
             return
         self._inited = True
         self.sp_world_size = size
-        rank, local_rank, world_size, local_world_size = get_dist_setting()
-        self.dp_world_size = world_size // size
-        self.device_mesh = init_device_mesh(
-            get_device().split(':')[0], mesh_shape=(world_size // size, size), mesh_dim_names=['data', 'sequence'])
+        self._init_device_mesh()
 
         from transformers.modeling_utils import ALL_ATTENTION_FUNCTIONS
         ALL_ATTENTION_FUNCTIONS['flash_attention_2_origin'] = ALL_ATTENTION_FUNCTIONS['flash_attention_2']
@@ -687,144 +468,8 @@ class Ulysses(SequenceParallel):
         self.model_dtype = next(model.parameters()).dtype
         self.tokenizer = tokenizer
 
-    def _pad_sp(self, tensor, padding_value, dim=-1):
-        # code borrowed from xtuner
-        length = tensor.shape[dim]
-        if length % self.sp_world_size == 0:
-            return tensor
-
-        pad_num = self.sp_world_size - (length % self.sp_world_size)
-        if not isinstance(padding_value, torch.Tensor):
-            # ids
-            pad_shape = ((*tensor.shape[:dim], pad_num, *tensor.shape[dim + 1:]) if dim != -1 else
-                         (*tensor.shape[:dim], pad_num))
-            pad = torch.full(pad_shape, padding_value, dtype=tensor.dtype, device=tensor.device)
-            tensor = torch.cat([tensor, pad], dim=dim)
-        else:
-            # For embeddings
-            tensor = torch.cat([tensor, padding_value.unsqueeze(0).repeat(tensor.shape[0], pad_num, 1)], dim=dim)
-        return tensor
-
-    def world_size(self):
-        return self.sp_world_size
-
-    def _split_sp(self, input, dim: int):
-        # code borrowed from xtuner
-        if self.sp_world_size == 1:
-            return input
-
-        rank = dist.get_rank(self.sp_group)
-        dim_size = input.size(dim)
-        assert dim_size % self.sp_world_size == 0, (f'The dimension to split ({dim_size}) is not a multiple of '
-                                                    f'world size ({self.sp_world_size}), cannot split tensor evenly')
-
-        tensor_list = torch.split(input, dim_size // self.sp_world_size, dim=dim)
-        output = tensor_list[rank].contiguous()
-
-        return output
-
-    def pad_and_split_inputs(self,
-                             input_ids,
-                             input_embeds,
-                             labels,
-                             position_ids,
-                             attention_mask,
-                             loss_scale,
-                             embed_tokens=None):
-        tokenizer = self.tokenizer
-        if input_ids is not None:
-            input_ids = self._pad_sp(input_ids, padding_value=tokenizer.pad_token_id, dim=-1)
-        if input_embeds is not None:
-            pad_emb = torch.zeros(
-                (1, embed_tokens.weight.shape[-1])).to(embed_tokens.weight.device).to(embed_tokens.weight.dtype)
-            input_embeds = self._pad_sp(input_embeds, padding_value=pad_emb, dim=1)
-        batch_size = input_ids.shape[
-            0] if input_ids is not None else input_embeds.shape[0] if input_embeds is not None else 1
-        if position_ids is not None:
-            position_ids = self._pad_sp(position_ids, padding_value=0, dim=-1)
-        if (input_ids is not None or input_embeds is not None) and batch_size > 1:
-            inputs = input_ids if input_ids is not None else input_embeds
-            attn_shape = inputs.shape[1]  # The sequence length
-            if attention_mask is None:
-                attention_mask = torch.ones_like(position_ids)
-            attention_mask = self._pad_sp(attention_mask, padding_value=0, dim=-1)
-            cache_position = torch.arange(0, attn_shape, device=inputs.device)
-            # pad attention mask to 4d to avoid calculation errors
-            attention_mask = self.causal_mask_func(attention_mask, inputs.to(self.model_dtype), cache_position, None,
-                                                   None)
-        if input_ids is not None:
-            input_ids = self._split_sp(input_ids, dim=1)
-        if input_embeds is not None:
-            input_embeds = self._split_sp(input_embeds, dim=1)
-        if position_ids is not None:
-            position_ids = self._split_sp(position_ids, dim=-1)
-        if labels is not None:
-            labels = self._pad_sp(labels, padding_value=-100, dim=-1)
-            labels = torch.roll(labels, shifts=-1, dims=-1)
-            labels = self._split_sp(labels, dim=-1)
-
-        if loss_scale is not None:
-            loss_scale = self._pad_sp(loss_scale, padding_value=0., dim=-1)
-            loss_scale = torch.roll(loss_scale, shifts=-1, dims=-1)
-            loss_scale = self._split_sp(loss_scale, dim=-1)
-
-        return input_ids, input_embeds, labels, position_ids, attention_mask, loss_scale
-
-    def reduce_outputs(self, loss, labels):
-        return loss
-
-    @property
-    def sp_rank(self):
-        return dist.get_rank(self.device_mesh['sequence'].get_group())
-
-    @property
-    def dp_rank(self):
-        return dist.get_rank(self.device_mesh['data'].get_group())
-
-    @property
-    def sp_group(self):
-        return self.device_mesh['sequence'].get_group()
-
-    @property
-    def dp_group(self):
-        return self.device_mesh['data'].get_group()
-
     def get_dataloader(self, trainer, dataset, batch_size):
-        data_collator = trainer.data_collator
-        if isinstance(dataset, datasets.Dataset):
-            dataset = trainer._remove_unused_columns(dataset, description='training')
-        else:
-            data_collator = trainer._get_collator_with_removed_columns(data_collator, description='training')
-        if hasattr(dataset, '__len__'):
-            sampler = UlyssesSampler(self, dataset, seed=42)
-            dataloader_params = {
-                'batch_size': batch_size,
-                'collate_fn': data_collator,
-                'num_workers': trainer.args.dataloader_num_workers,
-                'pin_memory': trainer.args.dataloader_pin_memory,
-                'persistent_workers': trainer.args.dataloader_persistent_workers,
-            }
-
-            if not isinstance(dataset, torch.utils.data.IterableDataset):
-                dataloader_params['sampler'] = sampler
-                dataloader_params['drop_last'] = trainer.args.dataloader_drop_last
-                dataloader_params['worker_init_fn'] = partial(
-                    seed_worker, num_workers=trainer.args.dataloader_num_workers, rank=trainer.args.process_index)
-
-            return DataLoaderShard(dataset, device=trainer.accelerator.device, **dataloader_params)
-        else:
-            dataloader_params = {
-                'collate_fn': data_collator,
-                'num_workers': trainer.args.dataloader_num_workers,
-                'pin_memory': trainer.args.dataloader_pin_memory,
-                'persistent_workers': trainer.args.dataloader_persistent_workers,
-                'prefetch_factor': trainer.args.dataloader_prefetch_factor
-            }
-            if dist.is_initialized() and dataloader_params['prefetch_factor']:
-                dataloader_params['prefetch_factor'] = dataloader_params['prefetch_factor'] * dist.get_world_size()
-            dataloader = DataLoader(dataset, batch_size=batch_size, **dataloader_params)
-            dataloader = UlyssesDispatcher(dataloader, self, trainer.accelerator.device)
-            return dataloader
+        return get_common_dataloader(self, trainer, dataset, batch_size, SequenceParallelSampler, SequenceParallelDispatcher)
 
     def prepare_trainer(self, trainer):
         # TODO hack methods, not cool
@@ -834,13 +479,13 @@ class Ulysses(SequenceParallel):
         trainer.ulysses = self
         if trainer.__class__.__name__ == 'Seq2SeqTrainer':
             trainer._origin_prepare_inputs = trainer._prepare_inputs
-            trainer._prepare_inputs = MethodType(partial(_prepare_inputs, ulysses=self), trainer)
-            trainer.compute_loss_func = partial(loss_scale_sp_func, ulysses=self)
+            trainer._prepare_inputs = MethodType(partial(_prepare_inputs, sp_instance=self), trainer)
+            trainer.compute_loss_func = partial(loss_scale_sp_func, sp_instance=self)
 
         elif trainer.__class__.__name__ == 'DPOTrainer':
             trainer._origin_prepare_inputs = trainer._prepare_inputs
-            trainer._prepare_inputs = MethodType(partial(_prepare_inputs, ulysses=self), trainer)
-            trainer.get_per_token_logps = partial(get_per_token_logps, ulysses=self)
+            trainer._prepare_inputs = MethodType(partial(_prepare_inputs, sp_instance=self), trainer)
+            trainer.get_per_token_logps = partial(get_per_token_logps, sp_instance=self)
 
         elif trainer.__class__.__name__ == 'GRPOTrainer':
             assert version.parse(trl.__version__) >= version.parse('0.18.0')
@@ -872,33 +517,4 @@ class Ulysses(SequenceParallel):
             trainer.get_origin_train_dataloader = trainer.get_train_dataloader
             trainer.get_train_dataloader = MethodType(get_train_dataloader, trainer)
 
-        from swift.plugin import metric
-        from swift.trainers import mixin
-        compute_acc_origin = metric.compute_acc
-
-        def compute_acc(preds, labels, *args, **kwargs) -> Dict[str, List[float]]:
-            # Gather preds and labels across the sp group
-            if isinstance(preds, np.ndarray):
-                preds = torch.from_numpy(preds).to(get_current_device())
-            if isinstance(labels, np.ndarray):
-                labels = torch.from_numpy(labels).to(get_current_device())
-            if labels.shape[1] > preds.shape[1]:
-                _, _, labels, _, _, _ = self.pad_and_split_inputs(None, None, labels, None, None, None)
-            shape0 = preds.shape[0]
-            preds_output = torch.empty((shape0 * self.sp_world_size, preds.shape[1]),
-                                       dtype=preds.dtype,
-                                       device=preds.device)
-            dist.all_gather_into_tensor(preds_output, preds, group=self.sp_group)
-            preds_output = torch.cat(preds_output.split(shape0, dim=0), dim=1)
-            shape0 = labels.shape[0]
-            labels_output = torch.empty((shape0 * self.sp_world_size, labels.shape[1]),
-                                        dtype=labels.dtype,
-                                        device=labels.device)
-            dist.all_gather_into_tensor(labels_output, labels, group=self.sp_group)
-            labels_output = torch.cat(labels_output.split(shape0, dim=0), dim=1)
-            # roll back to fit compute_acc
-            labels_output = torch.roll(labels_output, shifts=1, dims=1)
-            return compute_acc_origin(preds_output, labels_output, *args, **kwargs)
-
-        metric.compute_acc = compute_acc
-        mixin.compute_acc = compute_acc
+        setup_compute_acc(self)
